Tweak so building all examples doesn't explode :boom:
