+++ examples/universal/index.js
@@ -1,2 +1,1 @@
-require('babel/register');
-require('./server');
+require('./client');
